Modeling tabular data

Most statistical models require that data be represented as a Matrix-like collection of a single numeric type. Much of the data we want to model, however, is tabular data, where data is represented as a collection of fields with possibly heterogeneous types. One of the primary goals of StatsModels is to make it simpler to transform tabular data into matrix format suitable for statistical modeling.

At the moment, "tabular data" means an AbstractDataFrame. Ultimately, the goal is to support any tabular data format that adheres to a minimal API, regardless of backend.

The Formula type

The basic conceptual tool for this is the Formula, which has a left side and a right side, separated by ~:

julia> y ~ 1 + a
Formula: y ~ 1 + a

The left side of a formula conventionally represents dependent variables, and the right side independent variables (or regressors). Terms are separated by +. Basic terms are the integers 1 or 0—evaluated as the presence or absence of a constant intercept term, respectively—and variables like x, which will evaluate to the data source column with that name as a symbol (:x).

Individual variables can be combined into interaction terms with &, as in a&b, which will evaluate to the product of the columns named :a and :b. If either a or b are categorical, then the interaction term a&b generates all the product of each pair of the columns of a and b.

It's often convenient to include main effects and interactions for a number of variables. The * operator does this, expanding in the following way:

julia> Formula(StatsModels.Terms(y ~ 1 + a*b))
Formula: y ~ 1 + a + b + a & b

(We trigger parsing of the formula using the internal Terms type to show how the Formula expands).

This applies to higher-order interactions, too: a*b*c expands to the main effects, all two-way interactions, and the three way interaction a&b&c:

julia> Formula(StatsModels.Terms(y ~ 1 + a*b*c))
Formula: y ~ 1 + a + b + c + a & b + a & c + b & c + &(a,b,c)

Both the * and the & operators act like multiplication, and are distributive over addition:

julia> Formula(StatsModels.Terms(y ~ 1 + (a+b) & c))
Formula: y ~ 1 + c & a + c & b

julia> Formula(StatsModels.Terms(y ~ 1 + (a+b) * c))
Formula: y ~ 1 + a + b + c + c & a + c & b

The ModelFrame and ModelMatrix types

The main use of Formulas is for fitting statistical models based on tabular data. From the user's perspective, this is done by fit methods that take a Formula and a DataFrame instead of numeric matrices.

Internally, this is accomplished in three stages:

The Formula is parsed into Terms.

The Terms and the data source are wrapped in a ModelFrame.

A numeric ModelMatrix is generated from the ModelFrame and passed to the model's fit method.

StatsModels.ModelFrame — Type.

Wrapper which combines Formula (Terms) and an AbstractDataFrame

This wrapper encapsulates all the information that's required to transform data of the same structure as the wrapped data frame into a model matrix. This goes above and beyond what's expressed in the Formula itself, for instance including information on how each categorical variable should be coded.

Creating a ModelFrame first parses the Formula into Terms, checks which variables are categorical and determines the appropriate contrasts to use, and then creates the necessary contrasts matrices and stores the results.

Constructors

ModelFrame(f::Formula, df::AbstractDataFrame; contrasts::Dict = Dict())
ModelFrame(ex::Expr, d::AbstractDataFrame; contrasts::Dict = Dict())
ModelFrame(terms::Terms, df::AbstractDataFrame; contrasts::Dict = Dict())
# Inner constructors:
ModelFrame(df::AbstractDataFrame, terms::Terms, missing::BitArray)
ModelFrame(df::AbstractDataFrame, terms::Terms, missing::BitArray, contrasts::Dict{Symbol, ContrastsMatrix})

Arguments

f::Formula: Formula whose left hand side is the response and right hand side are the predictors.

df::AbstractDataFrame: The data being modeled. This is used at this stage to determine which variables are categorical, and otherwise held for ModelMatrix.

contrasts::Dict: An optional Dict of contrast codings for each categorical variable. Any unspecified variables will have DummyCoding. As a keyword argument, these can be either instances of a subtype of AbstractContrasts, or a ContrastsMatrix. For the inner constructor, they must be ContrastsMatrixes.

ex::Expr: An expression which will be converted into a Formula.

terms::Terms: For inner constructor, the parsed Terms from the Formula.

missing::BitArray: For inner constructor, indicates whether each row of df contains any missing data.

Examples

julia> df = DataFrame(x = 1:4, y = 5:9)
julia> mf = ModelFrame(y ~ 1 + x, df)

source

StatsModels.ModelMatrix — Type.

Convert a ModelFrame into a numeric matrix suitable for modeling

Constructors

ModelMatrix(mf::ModelFrame)
# Specify the type of the resulting matrix (default Matrix{Float64})
ModelMatrix{T <: AbstractFloatMatrix}(mf::ModelFrame)

source

StatsModels.Terms — Type.

Representation of parsed Formula

This is an internal type whose implementation is likely to change in the near future.

source